requirements.txt
Flask==2.0.1
requests==2.26.0
Werkzeug==2.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
